@@ -8,474 +8,300 @@ using System.IO;
 using System;
 using Microsoft.Extensions.Configuration;
 using Microsoft.AspNetCore.Identity;
-using OsmIntegrator.Roles;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Storage;
+using OsmIntegrator.Database.Models.Enums;
+using OsmIntegrator.Enums;
 
-namespace OsmIntegrator.Database.DataInitialization
+namespace OsmIntegrator.Database.DataInitialization;
+
+public class DataInitializer
 {
-  public class DataInitializer
+  private readonly double _overlapFactor;
+  private readonly int _zoomLevel;
+
+  public DataInitializer(IConfiguration configuration)
   {
-    private readonly double _overlapFactor;
-    private readonly int _zoomLevel;
+    _zoomLevel = int.Parse(configuration["ZoomLevel"]);
+    _overlapFactor = double.Parse(configuration["OverlapFactor"], NumberFormatInfo.InvariantInfo);
+  }
 
-    public DataInitializer(IConfiguration configuration)
+  public void Initialize(ApplicationDbContext db)
+  {
+    List<DbStop> gtfsStops = GetGtfsStopsList("Files/GtfsStops.txt");
+    List<DbStop> osmStops = GetOsmStopsList("Files/OsmStops.xml");
+
+    Initialize(db, gtfsStops, osmStops);
+  }
+
+  public void InitializeUsers(ApplicationDbContext db)
+  {
+    Guid editorRoleId = Guid.Parse("ff6cce18-706d-448c-89de-56212f1722ef");
+    Guid supervisorRoleId = Guid.Parse("3e4d73d6-6566-4518-a5b0-dce5a7016e24");
+    Guid coordinatorRoleId = Guid.Parse("17b91040-16cb-4fc6-9954-421cc8adc154");
+    Guid uploaderRoleId = Guid.Parse("f44243a2-7ea3-4327-9dbd-4ff97c164b33");
+    Guid adminRoleId = Guid.Parse("62f8bfff-d16a-4748-919c-56131148262e");
+
+    AddRole(db.Roles, editorRoleId, Roles.UserRoles.EDITOR);
+    AddRole(db.Roles, supervisorRoleId, Roles.UserRoles.SUPERVISOR);
+    AddRole(db.Roles, coordinatorRoleId, Roles.UserRoles.COORDINATOR);
+    AddRole(db.Roles, uploaderRoleId, Roles.UserRoles.UPLOADER);
+    AddRole(db.Roles, adminRoleId, Roles.UserRoles.ADMIN);
+
+    AddUser(db.Users, db.UserRoles, Guid.Parse("c514ae13-d80a-40b1-90d0-df88ccca73ec"), 
+      "user1@abcd.pl", new List<Guid>() { editorRoleId });
+    AddUser(db.Users, db.UserRoles, Guid.Parse("9815ba57-e990-4f91-a8e9-17abe977d681"), 
+      "editor1@abcd.pl", new List<Guid>() { editorRoleId });
+    AddUser(db.Users, db.UserRoles, Guid.Parse("2afa8c7d-3b29-4d33-b69c-afc71626b109"), 
+      "supervisor1@abcd.pl", new List<Guid>() { supervisorRoleId });
+    AddUser(db.Users, db.UserRoles, Guid.Parse("58fb0db2-0d67-4aa4-af63-9e7111d0b346"), 
+      "supervisor2@abcd.pl", new List<Guid>() { supervisorRoleId, editorRoleId });
+    AddUser(db.Users, db.UserRoles, Guid.Parse("cbfbbb17-eaed-46e8-b972-8c9fd0f8fa5b"), 
+      "coordinator1@abcd.pl", new List<Guid>() { coordinatorRoleId });
+    AddUser(db.Users, db.UserRoles, Guid.Parse("55529e52-ba92-4727-94bd-53bcc1be06c8"), 
+      "admin@abcd.pl", new List<Guid>() { adminRoleId });
+    AddUser(db.Users, db.UserRoles, Guid.Parse("55529e52-ba92-4727-94bd-53bcc1be06c7"), 
+      "test@rozwiazaniadlaniewidomych.org", new List<Guid>() { editorRoleId, adminRoleId, supervisorRoleId });
+    AddUser(db.Users, db.UserRoles, Guid.Parse("a658f1c1-e91f-4bde-afb2-58c50b0d170a"), 
+      "editor2@abcd.pl", new List<Guid>() { editorRoleId });
+    
+    db.SaveChanges();
+  }
+
+  public void InitializeStopsAndTiles(
+    ApplicationDbContext db, List<DbStop> gtfsStops, List<DbStop> osmStops)
+  {
+    gtfsStops ??= new();
+    osmStops ??= new();
+
+    List<DbStop> allStops = new List<DbStop>();
+    allStops.AddRange(osmStops);
+    allStops.AddRange(gtfsStops);
+    List<DbTile> tiles = GetTiles(allStops);
+
+    db.Stops.AddRange(allStops);
+    db.Tiles.AddRange(tiles);
+
+    InitializeOsmConnections(db, gtfsStops, osmStops);
+
+    db.SaveChanges();
+  }
+
+  private void InitializeOsmConnections(ApplicationDbContext db, List<DbStop> gtfsStops, List<DbStop> osmStops)
+  {
+    if(gtfsStops == null || osmStops == null) return;
+
+    ApplicationUser supervisor = db.Users.First(x => x.UserName == "supervisor2");
+
+    List<DbConnection> connections = new();
+
+    foreach (DbStop osmStop in osmStops)
     {
-      _zoomLevel = int.Parse(configuration["ZoomLevel"]);
-      _overlapFactor = double.Parse(configuration["OverlapFactor"], NumberFormatInfo.InvariantInfo);
-    }
-
-    public void Initialize(ApplicationDbContext db)
-    {
-      List<DbStop> gtfsStops = GetGtfsStopsList();
-      List<DbStop> osmStops = GetOsmStopsList();
-
-      Initialize(db, gtfsStops, osmStops);
-      MigrateNotes(db);
-      MigrateTileUsers(db);
-    }
-
-    public void Initialize(ApplicationDbContext db, List<DbStop> gtfsStops, List<DbStop> osmStops)
-    {
-      using (var transaction = db.Database.BeginTransaction())
+      if (long.TryParse(osmStop.Ref, out long @ref))
       {
-        try
+        DbStop gtfsStop = gtfsStops.FirstOrDefault(x => x.StopId == @ref);
+
+        if (gtfsStop != null)
         {
-          if (db.Stops.Count() > 0)
+          connections.Add(new DbConnection()
           {
-            return;
-          }
-
-          Guid editorRoleId = Guid.Parse("ff6cce18-706d-448c-89de-56212f1722ef");
-          Guid supervisorRoleId = Guid.Parse("3e4d73d6-6566-4518-a5b0-dce5a7016e24");
-          Guid coordinatorRoleId = Guid.Parse("17b91040-16cb-4fc6-9954-421cc8adc154");
-          Guid uploaderRoleId = Guid.Parse("f44243a2-7ea3-4327-9dbd-4ff97c164b33");
-          Guid adminRoleId = Guid.Parse("62f8bfff-d16a-4748-919c-56131148262e");
-
-          AddRole(db.Roles, editorRoleId, OsmIntegrator.Roles.UserRoles.EDITOR);
-          AddRole(db.Roles, supervisorRoleId, OsmIntegrator.Roles.UserRoles.SUPERVISOR);
-          AddRole(db.Roles, coordinatorRoleId, OsmIntegrator.Roles.UserRoles.COORDINATOR);
-          AddRole(db.Roles, uploaderRoleId, OsmIntegrator.Roles.UserRoles.UPLOADER);
-          AddRole(db.Roles, adminRoleId, OsmIntegrator.Roles.UserRoles.ADMIN);
-
-          AddUser(db.Users, db.UserRoles, Guid.Parse("c514ae13-d80a-40b1-90d0-df88ccca73ec"), "user1@abcd.pl", new List<Guid>() { editorRoleId });
-          AddUser(db.Users, db.UserRoles, Guid.Parse("9815ba57-e990-4f91-a8e9-17abe977d681"), "editor1@abcd.pl", new List<Guid>() { editorRoleId });
-          AddUser(db.Users, db.UserRoles, Guid.Parse("2afa8c7d-3b29-4d33-b69c-afc71626b109"), "supervisor1@abcd.pl", new List<Guid>() { supervisorRoleId });
-          AddUser(db.Users, db.UserRoles, Guid.Parse("58fb0db2-0d67-4aa4-af63-9e7111d0b346"), "supervisor2@abcd.pl", new List<Guid>() { supervisorRoleId, editorRoleId });
-          AddUser(db.Users, db.UserRoles, Guid.Parse("cbfbbb17-eaed-46e8-b972-8c9fd0f8fa5b"), "coordinator1@abcd.pl", new List<Guid>() { coordinatorRoleId });
-          AddUser(db.Users, db.UserRoles, Guid.Parse("55529e52-ba92-4727-94bd-53bcc1be06c8"), "admin@abcd.pl", new List<Guid>() { adminRoleId });
-          AddUser(db.Users, db.UserRoles, Guid.Parse("55529e52-ba92-4727-94bd-53bcc1be06c7"), "test@rozwiazaniadlaniewidomych.org", new List<Guid>() { editorRoleId, adminRoleId, supervisorRoleId });
-          AddUser(db.Users, db.UserRoles, Guid.Parse("a658f1c1-e91f-4bde-afb2-58c50b0d170a"), "editor2@abcd.pl", new List<Guid>() { editorRoleId });
-          db.SaveChanges();
-
-          List<DbStop> allStops = new List<DbStop>();
-          allStops.AddRange(osmStops);
-          allStops.AddRange(gtfsStops);
-          List<DbTile> tiles = GetTiles(allStops);
-
-          DbTile tile_2263_1385 = tiles.FirstOrDefault(x => x.X == 2263 && x.Y == 1385);
-
-          List<ApplicationUser> users = db.Users.Where(x => x.UserName.Contains("editor")).ToList();
-          ApplicationUser editor1 = users[0];
-          ApplicationUser editor2 = users[1];
-
-          List<ApplicationUser> supervisors = db.Users.Where(x => x.UserName.Contains("supervisor")).ToList();
-          ApplicationUser supervisor1 = supervisors[0];
-
-          tiles[0].Users = tiles[1].Users = tile_2263_1385.Users = new List<ApplicationUser> { editor1 };
-          tiles[0].EditorApproved = tiles[1].EditorApproved = tile_2263_1385.EditorApproved = editor1;
-          tiles[0].EditorApprovalTime = tiles[1].EditorApprovalTime = tile_2263_1385.EditorApprovalTime = DateTime.Now;
-
-          db.Stops.AddRange(allStops);
-          db.Tiles.AddRange(tiles);
-          db.SaveChanges();
-
-          DbTile tile = db.Tiles.First(x => x.X == 2263 && x.Y == 1385);
-
-          DbNote note1 = new DbNote
-          {
-            Text = "Not approved",
-            Lat = tile.MinLat,
-            Lon = tile.MinLon,
-            UserId = editor1.Id,
-            Status = NoteStatus.Created,
-            TileId = tile.Id
-          };
-
-          DbNote note2 = new DbNote
-          {
-            Text = "Approved",
-            Lat = tile.MaxLat,
-            Lon = tile.MinLon,
-            UserId = editor1.Id,
-            Status = NoteStatus.Approved,
-            Approver = supervisor1,
-            TileId = tile.Id
-          };
-
-          DbNote note3 = new DbNote
-          {
-            Text = "Supervisor approved",
-            Lat = tile.MaxLat,
-            Lon = tile.MinLon,
-            UserId = supervisor1.Id,
-            Status = NoteStatus.Rejected,
-            Approver = supervisor1,
-            TileId = tile.Id
-          };
-
-          DbNote note4 = new DbNote
-          {
-            Text = "Supervisor not approved",
-            Lat = tile.MaxLat,
-            Lon = tile.MinLon,
-            UserId = supervisor1.Id,
-            Status = NoteStatus.Created,
-            TileId = tile.Id
-          };
-
-          DbNote note5 = new DbNote
-          {
-            Text = "Editor 2",
-            Lat = tile.MaxLat,
-            Lon = tile.MinLon,
-            UserId = editor2.Id,
-            Status = NoteStatus.Created,
-            TileId = tile.Id
-          };
-
-          db.Notes.Add(note1);
-          db.Notes.Add(note2);
-          db.Notes.Add(note3);
-          db.Notes.Add(note4);
-          db.Notes.Add(note5);
-          db.SaveChanges();
-
-          List<DbConnections> connections = new List<DbConnections>();
-
-          foreach (DbStop osmStop in osmStops)
-          {
-            DbStop gtfsStop = gtfsStops.FirstOrDefault(x => x.StopId == osmStop.Ref);
-            if (gtfsStop != null)
-            {
-              connections.Add(new DbConnections()
-              {
-                GtfsStop = gtfsStop,
-                OsmStop = osmStop,
-                Imported = true
-              });
-            }
-          }
-
-          db.Connections.AddRange(connections);
-          db.SaveChanges();
-
-          transaction.Commit();
-        }
-        catch (Exception e)
-        {
-          Console.WriteLine(e);
-          transaction.Rollback();
-          throw;
+            GtfsStop = gtfsStop,
+            OsmStop = osmStop,
+            UserId = supervisor.Id,
+            OperationType = ConnectionOperationType.Added
+          });
         }
       }
     }
 
-    public void MigrateNotes(ApplicationDbContext db)
+    db.Connections.AddRange(connections);
+    db.SaveChanges();
+  }
+
+  private void Initialize(ApplicationDbContext db, List<DbStop> gtfsStops, List<DbStop> osmStops)
+  {
+    using IDbContextTransaction transaction = db.Database.BeginTransaction();
+    try
     {
-      using (var transaction = db.Database.BeginTransaction())
+      if (db.Stops.Any())
       {
-        if (db.Conversations.Count() > 0)
-        {
-          return;
-        }
-
-        try
-        {
-          foreach (DbNote note in db.Notes)
-          {
-            DbConversation conversation = new DbConversation()
-            {
-              Id = new Guid(),
-              Lat = note.Lat,
-              Lon = note.Lon,
-              TileId = note.TileId,
-            };
-            db.Conversations.Add(conversation);
-
-            db.Messages.Add(new DbMessage()
-            {
-              Id = new Guid(),
-              ConversationId = conversation.Id,
-              Status = NoteStatus.Created,
-              Text = note.Text,
-              UserId = note.UserId,
-              CreatedAt = DateTime.Now.AddMinutes(-5),
-            });
-
-            if (note.Status != NoteStatus.Created)
-            {
-              db.Messages.Add(new DbMessage()
-              {
-                Id = new Guid(),
-                ConversationId = conversation.Id,
-                Status = note.Status,
-                UserId = note.UserId,
-                CreatedAt = DateTime.Now,
-              });
-            }
-          }
-          db.SaveChanges();
-          transaction.Commit();
-        }
-        catch (Exception e)
-        {
-          Console.WriteLine(e);
-          transaction.Rollback();
-          throw;
-        }
+        return;
       }
+
+      InitializeUsers(db);
+      InitializeStopsAndTiles(db, gtfsStops, osmStops);
+      InitializeOsmConnections(db, gtfsStops, osmStops);
+
+      transaction.Commit();
     }
-    public void MigrateTileUsers(ApplicationDbContext db)
+    catch (Exception e)
     {
-      using (var transaction = db.Database.BeginTransaction())
-      {
-        if (db.TileUsers.Count() > 0)
-        {
-          return;
-        }
-        try
-        {
-          ApplicationUser supervisor = db.Users.First(x => x.UserName == "supervisor1");
-
-          foreach (ApplicationUser user in db.Users.Include(x => x.Tiles).ToList())
-          {
-            foreach (DbTile tile in user.Tiles)
-            {
-              db.TileUsers.Add(new DbTileUser()
-              {
-                Id = new Guid(),
-                User = user,
-                Tile = tile,
-                Role = db.Roles.Where(x => x.Name == UserRoles.EDITOR).First()
-              });
-
-              db.TileUsers.Add(new DbTileUser()
-              {
-                Id = new Guid(),
-                User = supervisor,
-                Tile = tile,
-                Role = db.Roles.Where(x => x.Name == UserRoles.SUPERVISOR).First()
-              });
-            }
-          }
-
-          db.SaveChanges();
-          transaction.Commit();
-        }
-        catch (Exception e)
-        {
-          Console.WriteLine(e);
-          transaction.Rollback();
-          throw;
-        }
-      }
+      Console.WriteLine(e);
+      transaction.Rollback();
+      throw;
     }
-    public void AddRole(DbSet<ApplicationRole> roles, Guid id, string name)
+  }
+
+  private void AddRole(DbSet<ApplicationRole> roles, Guid id, string name)
+  {
+    roles.Add(new ApplicationRole()
     {
-      roles.Add(new ApplicationRole()
+      Id = id,
+      Name = name,
+      NormalizedName = name.ToUpper()
+    });
+  }
+
+  private void AddUser(
+    DbSet<ApplicationUser> users,
+    DbSet<IdentityUserRole<Guid>> userRoles,
+    Guid id,
+    string email,
+    List<Guid> roleIds = null)
+  {
+    var hasher = new PasswordHasher<ApplicationUser>();
+    string name = email.Split("@")[0];
+    users.Add(new ApplicationUser
+    {
+      Id = id,
+      UserName = name,
+      NormalizedUserName = name.ToUpper(),
+      Email = email,
+      NormalizedEmail = email.ToUpper(),
+      EmailConfirmed = true,
+      PasswordHash = hasher.HashPassword(null, $"{name}#12345678"),
+      SecurityStamp = string.Empty
+    });
+
+    if (roleIds == null) return;
+
+    foreach (Guid roleId in roleIds)
+    {
+      userRoles.Add(new IdentityUserRole<Guid>
       {
-        Id = id,
-        Name = name,
-        NormalizedName = name.ToUpper()
+        RoleId = roleId,
+        UserId = id
       });
     }
+  }
 
-    public void AddUser(
-        DbSet<ApplicationUser> users,
-        DbSet<IdentityUserRole<Guid>> userRoles,
-        Guid id,
-        string email,
-        List<Guid> roleIds = null)
+  public static List<DbStop> GetGtfsStopsList(string fileName)
+  {
+    List<string[]> csvStopList = CsvParser.Parse(fileName);
+    List<DbStop> ztmStopList = csvStopList.Select((x, _) => new DbStop()
     {
-      var hasher = new PasswordHasher<ApplicationUser>();
-      string name = email.Split("@")[0];
-      users.Add(new ApplicationUser
-      {
-        Id = id,
-        UserName = name,
-        NormalizedUserName = name.ToUpper(),
-        Email = email,
-        NormalizedEmail = email.ToUpper(),
-        EmailConfirmed = true,
-        PasswordHash = hasher.HashPassword(null, $"{name}#12345678"),
-        SecurityStamp = string.Empty
-      });
+      Id = Guid.NewGuid(),
+      StopId = long.Parse(x[0]),
+      Number = x[1],
+      Name = x[2],
+      Lat = double.Parse(x[4], CultureInfo.InvariantCulture.NumberFormat),
+      Lon = double.Parse(x[5], CultureInfo.InvariantCulture.NumberFormat),
+      StopType = StopType.Gtfs,
+      ProviderType = ProviderType.Ztm
+    }).ToList();
+    return ztmStopList;
+  }
 
-      if (roleIds == null) return;
+  public List<DbStop> GetOsmStopsList(string fileName)
+  {
+    List<DbStop> result = new();
 
-      foreach (Guid roleId in roleIds)
+    XmlSerializer serializer = new(typeof(Osm));
+
+    using (Stream reader = new FileStream(fileName, FileMode.Open))
+    {
+      Osm osmRoot = (Osm)serializer.Deserialize(reader);
+
+      foreach (Node node in osmRoot!.Node)
       {
-        userRoles.Add(new IdentityUserRole<Guid>
+        DbStop stop = new()
         {
-          RoleId = roleId,
-          UserId = id
-        });
+          Id = Guid.NewGuid(),
+          StopId = long.Parse(node.Id),
+          Lat = double.Parse(node.Lat, CultureInfo.InvariantCulture),
+          Lon = double.Parse(node.Lon, CultureInfo.InvariantCulture),
+          StopType = StopType.Osm,
+          ProviderType = ProviderType.Ztm,
+          Version = node.Version,
+          Changeset = node.Changeset
+        };
+
+        List<Models.JsonFields.Tag> tempTags = new();
+
+        node.Tag.ForEach(x => tempTags.Add(new Models.JsonFields.Tag()
+        {
+          Key = x.K,
+          Value = x.V
+        }));
+        stop.Tags = tempTags;
+
+        var nameTag = tempTags.FirstOrDefault(x => x.Key.ToLower() == Constants.NAME);
+        stop.Name = nameTag?.Value;
+
+        var refTag = tempTags.FirstOrDefault(x => x.Key.ToLower() == Constants.REF);
+        stop.Ref = refTag?.Value;
+
+        var localRefTag = tempTags.FirstOrDefault(x => x.Key.ToLower() == Constants.LOCAL_REF);
+        stop.Number = localRefTag?.Value;
+
+        result.Add(stop);
       }
     }
 
-    public List<DbStop> GetGtfsStopsList()
+    return result;
+  }
+
+  private List<DbTile> GetTiles(List<DbStop> stops)
+  {
+    Dictionary<Point<long>, DbTile> result = new();
+
+    foreach (DbStop stop in stops)
     {
-      List<string[]> csvStopList = CsvParser.Parse("Files/GtfsStops.txt");
-      List<DbStop> ztmStopList = csvStopList.Select((x, index) => new DbStop()
+      Point<long> tileCoordinates = TilesHelper.WorldToTilePos(stop.Lon, stop.Lat, _zoomLevel);
+
+      if (result.ContainsKey(tileCoordinates))
       {
-        Id = Guid.NewGuid(),
-        StopId = long.Parse(x[0]),
-        Number = x[1],
-        Name = x[2],
-        Lat = double.Parse(x[4], CultureInfo.InvariantCulture.NumberFormat),
-        Lon = double.Parse(x[5], CultureInfo.InvariantCulture.NumberFormat),
-        StopType = StopType.Gtfs,
-        ProviderType = ProviderType.Ztm
-      }).ToList();
-      return ztmStopList;
-    }
+        DbTile existingTile = result[tileCoordinates];
+        stop.TileId = existingTile.Id;
 
-    public List<DbStop> GetOsmStopsList()
-    {
-      List<DbStop> result = new List<DbStop>();
-
-      XmlSerializer serializer = new XmlSerializer(typeof(Osm));
-
-      using (Stream reader = new FileStream("Files/OsmStops.xml", FileMode.Open))
-      {
-        Osm osmRoot = (Osm)serializer.Deserialize(reader);
-
-        foreach (Node node in osmRoot.Node)
-        {
-          DbStop stop = new DbStop
-          {
-            Id = Guid.NewGuid(),
-            StopId = long.Parse(node.Id),
-            Lat = double.Parse(node.Lat, CultureInfo.InvariantCulture),
-            Lon = double.Parse(node.Lon, CultureInfo.InvariantCulture),
-            StopType = StopType.Osm,
-            ProviderType = ProviderType.Ztm,
-            Version = node.Version,
-            Changeset = node.Changeset
-          };
-
-          List<Models.Tag> tempTags = new List<Models.Tag>();
-
-          node.Tag.ForEach(x => tempTags.Add(new Models.Tag()
-          {
-            Key = x.K,
-            Value = x.V
-          }));
-          stop.Tags = tempTags;
-
-          var nameTag = tempTags.FirstOrDefault(x => x.Key.ToLower() == "name");
-          stop.Name = nameTag?.Value;
-
-          var refTag = tempTags.FirstOrDefault(x => x.Key.ToLower() == "ref");
-          long refVal;
-          if (refTag != null && long.TryParse(refTag.Value, out refVal))
-          {
-            stop.Ref = refVal;
-          }
-
-          var localRefTag = tempTags.FirstOrDefault(x => x.Key.ToLower() == "local_ref");
-          if (localRefTag != null && !string.IsNullOrEmpty(localRefTag.Value))
-          {
-            stop.Number = localRefTag.Value;
-          }
-
-          result.Add(stop);
-        }
+        continue;
       }
 
-      return result;
+      Point<double> leftUpperCorner = TilesHelper.TileToWorldPos(
+        tileCoordinates.X, tileCoordinates.Y, _zoomLevel
+      );
+
+      Point<double> rightBottomCorner = TilesHelper.TileToWorldPos(
+        tileCoordinates.X + 1, tileCoordinates.Y + 1, _zoomLevel
+      );
+
+      DbTile newTile = new(tileCoordinates.X, tileCoordinates.Y,
+        leftUpperCorner.X, rightBottomCorner.X,
+        rightBottomCorner.Y, leftUpperCorner.Y,
+        _overlapFactor);
+
+      stop.TileId = newTile.Id;
+      result.Add(tileCoordinates, newTile);
     }
 
-    public List<DbTile> GetTiles(List<DbStop> stops)
-    {
-      Dictionary<Point<long>, DbTile> result = new Dictionary<Point<long>, DbTile>();
+    return result.Values.ToList();
+  }
 
-      foreach (DbStop stop in stops)
-      {
-        Point<long> tileXY = TilesHelper.WorldToTilePos(stop.Lon, stop.Lat, _zoomLevel);
+  public void ClearDatabase(ApplicationDbContext db)
+  {
+    db.Connections.RemoveRange(db.Connections);
+    db.Stops.RemoveRange(db.Stops);
+    db.Tiles.RemoveRange(db.Tiles);
 
-        if (result.ContainsKey(tileXY))
-        {
-          DbTile existingTile = result[tileXY];
-          stop.TileId = existingTile.Id;
+    db.Conversations.RemoveRange(db.Conversations);
+    db.Messages.RemoveRange(db.Messages);
 
-          if (stop.StopType == StopType.Gtfs)
-          {
-            existingTile.GtfsStopsCount++;
-          }
-          else
-          {
-            existingTile.OsmStopsCount++;
-          }
+    db.Roles.RemoveRange(db.Roles);
+    db.Users.RemoveRange(db.Users);
+    db.UserRoles.RemoveRange(db.UserRoles);
 
-          continue;
-        }
+    db.ChangeReports.RemoveRange(db.ChangeReports);
 
-        Point<double> leftUpperCorner = TilesHelper.TileToWorldPos(
-            tileXY.X, tileXY.Y, _zoomLevel
-        );
-
-        Point<double> rightBottomCorner = TilesHelper.TileToWorldPos(
-            tileXY.X + 1, tileXY.Y + 1, _zoomLevel
-        );
-
-        DbTile newTile = new DbTile(tileXY.X, tileXY.Y,
-            leftUpperCorner.X, rightBottomCorner.X,
-            rightBottomCorner.Y, leftUpperCorner.Y,
-            _overlapFactor);
-
-        if (stop.StopType == StopType.Gtfs)
-        {
-          newTile.GtfsStopsCount++;
-        }
-        else
-        {
-          newTile.OsmStopsCount++;
-        }
-
-        stop.TileId = newTile.Id;
-        result.Add(tileXY, newTile);
-      }
-
-      return result.Values.ToList();
-    }
-
-    public void ClearDatabase(ApplicationDbContext db)
-    {
-      using (var transaction = db.Database.BeginTransaction())
-      {
-        try
-        {
-          db.Notes.RemoveRange(db.Notes);
-          db.Connections.RemoveRange(db.Connections);
-          db.Stops.RemoveRange(db.Stops);
-          db.Tiles.RemoveRange(db.Tiles);
-
-          db.Roles.RemoveRange(db.Roles);
-          db.Users.RemoveRange(db.Users);
-          db.UserRoles.RemoveRange(db.UserRoles);
-
-          db.SaveChanges();
-
-          transaction.Commit();
-        }
-        catch (Exception e)
-        {
-          Console.WriteLine(e);
-          transaction.Rollback();
-          throw;
-        }
-      }
-    }
+    db.SaveChanges();
   }
 }
