@@ -1,11 +1,9 @@
 using System;
-using System.Collections.Generic;
 using Microsoft.AspNetCore.Identity;
 
-namespace OsmIntegrator.Database.Models
-{
-    public class ApplicationRole : IdentityRole<Guid>
-    {
+namespace OsmIntegrator.Database.Models;
 
-    }
+public class ApplicationRole : IdentityRole<Guid>
+{
+
 }
