@@ -1,80 +1,75 @@
-﻿using Newtonsoft.Json;
+﻿using Microsoft.EntityFrameworkCore;
+using Newtonsoft.Json;
 using OsmIntegrator.ApiModels;
-using OsmIntegrator.ApiModels.Auth;
+using OsmIntegrator.ApiModels.Connections;
+using OsmIntegrator.Database;
 using OsmIntegrator.Database.Models;
-using OsmIntegrator.Tests.Helpers.Base;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Net.Http;
-using System.Net.Http.Headers;
 using System.Text;
 using System.Threading.Tasks;
 
 namespace OsmIntegrator.Tests.Helpers
 {
-    public class ConnectionHelper : BaseHelper
+  public class ConnectionHelper
+  {
+    private readonly ApplicationDbContext _dbContext;
+
+    private readonly Dictionary<int, int> _stops = new()
     {
-        private const string Route = "/api/Connections";
-        private LoginData _defaultLoginData = new LoginData
-        {
-            Email = "supervisor1@abcd.pl",
-            Password = "supervisor1#12345678",
-        };
+      { 159541, 1 },      // [1]  Tile: 0 = LEFT    GTFS:   Stara Ligota Rolna 1
+      { 1831941739, 0},   // [2]  Tile: 0 = LEFT    OSM:    Stara Ligota Rolna
+      { 159542, 1 },      // [3]  Tile: 0 = LEFT    GTFS:   Stara Ligota Rolna 2  In margin of the right tile
+      { 1905028012, 0 },  // [4]  Tile: 0 = LEFT    OSM:    Stara Ligota Rolna    In margin of the right tile
+      { 159077, 1 },      // [5]  Tile: 1 = RIGHT   GTFS:   Brynów Orkana 2
+      { 1831944331, 0 },  // [6]  Tile: 1 = RIGHT   OSM:    Brynów Orkana
+      { 1905039171, 0 },  // [7]  Tile: 1 = RIGHT   OSM:    Brynów Orkana
+      { 159076, 1 },      // [8]  Tile: 1 = RIGHT   GTFS:   Brynów Orkana 1
+      { 159061, 1 },      // [9]  Tile: 1 = RIGHT   GTFS:   Brynów Dworska 1
+      { 1584594015, 0}    // [10] Tile: 1 = RIGHT   OSM:    Brynów Dworska
+    };
 
-        public NewConnectionAction CreateConnection(int gtfsStopId, int osmStopId, int tileId)
-        { 
+    private Dictionary<int, DbStop> Stops { get; }
+    public List<DbTile> Tiles {get; }
 
-            var stopHelper = new StopHelper(_client, _defaultLoginData);
-            var stopDict = stopHelper.GetTestStopDict();
-            var tileHelper = new TileHelper(_client, _defaultLoginData);
-            var tileDict = tileHelper.GetTestTileDict();
+    public ConnectionHelper(ApplicationDbContext dbContext)
+    {
+      _dbContext = dbContext;
 
-            var connectionAction = new NewConnectionAction
-            {
-                OsmStopId = stopDict[osmStopId].Id,
-                GtfsStopId = stopDict[gtfsStopId].Id,
-                TileId = tileDict[tileId].Id,
-            };
-
-            return connectionAction;
-        }
-
-        public async Task<List<DbConnections>> GetConnectionListAsync()
-        {
-            var response = await _client.GetAsync(Route);
-            var jsonResponse = await response.Content.ReadAsStringAsync();
-            var list = JsonConvert.DeserializeObject<DbConnections[]>(jsonResponse).Where(f => f.OperationType==Enums.ConnectionOperationType.Added).ToList();
-
-            return list;
-        }
-
-        public async Task<HttpResponseMessage> CreateConnection(NewConnectionAction connectionAction)
-        {
-            var jsonConnectionAction = JsonConvert.SerializeObject(connectionAction);
-            var content = new StringContent(jsonConnectionAction, Encoding.UTF8, "application/json");
-            var response = await _client.PutAsync(Route, content);
-            return response;
-        }
-
-        public async Task<HttpResponseMessage> DeleteConnection(ConnectionAction connectionAction)
-        {
-            var jsonConnectionAction = JsonConvert.SerializeObject(connectionAction);
-            var content = new StringContent(jsonConnectionAction, Encoding.UTF8, "application/json");
-
-            HttpRequestMessage requestMessage = new HttpRequestMessage
-            {
-                Method = HttpMethod.Delete,
-                RequestUri = new Uri($"{HttpAddressAndPort}{Route}"),
-                Content = content,
-            };
-            var response = await _client.SendAsync(requestMessage);
-            return response;
-        }
-
-
-        public ConnectionHelper(HttpClient factoryClient, LoginData loginData) : base(factoryClient, loginData)
-        {
-        }
+      Stops = GetStopsDict();
+      Tiles = new List<DbTile>{
+        Stops[1].Tile, Stops[9].Tile
+      };
     }
+
+    private Dictionary<int, DbStop> GetStopsDict()
+    {
+      Dictionary<int, DbStop> result = new();
+
+      List<DbStop> stops = _dbContext.Stops.Include(x => x.Tile).ToList();
+
+      int index = 1;
+      foreach ((int key, int value) in _stops)
+      {
+        result.Add(index, stops.First(x => x.StopId == key && (int)x.StopType == value));
+        index++;
+      }
+
+      return result;
+    }
+
+    public NewConnectionAction CreateConnection(int gtfsStopId, int osmStopId, int tileId)
+    {
+      NewConnectionAction connectionAction = new()
+      {
+        OsmStopId = Stops[osmStopId].Id,
+        GtfsStopId = Stops[gtfsStopId].Id,
+        TileId = Tiles[tileId].Id,
+      };
+
+      return connectionAction;
+    }
+  }
 }
