@@ -1,7 +1,7 @@
 using System;
 using System.Linq;
 using AutoMapper;
-using OsmIntegrator.ApiModels;
+using OsmIntegrator.ApiModels.Conversation;
 using OsmIntegrator.Database.Models;
 
 namespace OsmIntegrator.AutoMapper
