@@ -1,18 +1,22 @@
+using System.Globalization;
 using AutoMapper;
-using OsmIntegrator.ApiModels;
+using Microsoft.Extensions.Configuration;
+using OsmIntegrator.ApiModels.Tiles;
 using OsmIntegrator.Database.Models;
 
 namespace OsmIntegrator.AutoMapper
 {
-    public class TileProfile : Profile
+  public class TileProfile : Profile
+  {
+    public TileProfile(IConfiguration configuration)
     {
-        public TileProfile()
-        {
-            AllowNullCollections = true;
-            CreateMap<DbTile, Tile>()
-                .ForMember(x => x.UsersCount, o => o.MapFrom(x => x.TileUsers.Count))
-                .ForMember(x => x.ApprovedBySupervisor, o => o.MapFrom(x => x.SupervisorApprovedId != null))
-                .ForMember(x => x.ApprovedByEditor, o => o.MapFrom(x => x.EditorApprovedId != null));
-        }
+      byte zoomLevel = byte.Parse(configuration["ZoomLevel"], NumberFormatInfo.InvariantInfo);
+
+      AllowNullCollections = true;
+      CreateMap<DbTile, Tile>()
+          .ForMember(x => x.ZoomLevel, o => o.MapFrom(x => zoomLevel))
+          .ForMember(x => x.AssignedUserName, 
+            o => o.MapFrom(x => x.AssignedUser == null ? string.Empty : x.AssignedUser.UserName));
     }
+  }
 }
