@@ -3,6 +3,6 @@ using OsmIntegrator.Interfaces;
 
 namespace OsmIntegrator.Interfaces {
     public interface IUseCase<in TInput> {
-        public Task<AUseCaseResponse> Handle(TInput useCaseData);
+        Task<AUseCaseResponse> Handle(TInput useCaseData);
     }
 }
